[MOVE](server): community 관련 service 코드 디렉터리 위치 이동
